requirements.txt
yfinance
python-telegram-bot
schedule
matplotlib
pillow
